Rename `getWorkingDir` to `parsePath`
